pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "solana-mcp"
version = "0.2.0"
description = "RAG-powered MCP server for Solana runtime and SIMDs"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "be.nvy" }
]
keywords = ["solana", "mcp", "rag", "runtime", "simd", "search", "blockchain"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.0.0",
    "lancedb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "click>=8.0.0",
    "tree-sitter>=0.20.0",
    "tree-sitter-rust>=0.20.0",
    "tree-sitter-c>=0.20.0",
    "gitpython>=3.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
voyage = [
    "voyageai>=0.2.0",
]

[project.urls]
Homepage = "https://github.com/be-nvy/solana-mcp"
Documentation = "https://github.com/be-nvy/solana-mcp#readme"
Repository = "https://github.com/be-nvy/solana-mcp.git"
Issues = "https://github.com/be-nvy/solana-mcp/issues"

[project.scripts]
solana-mcp = "solana_mcp.cli:main"
sol-mcp = "solana_mcp.server:run"

[tool.hatch.build.targets.wheel]
packages = ["src/solana_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
